requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Werkzeug==3.0.1
SQLAlchemy==2.0.27
WTForms==3.1.2
email-validator==2.1.0.post1
Pillow==10.1.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
alembic==1.13.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
click==8.1.7
blinker==1.7.0
greenlet==3.0.3
typing_extensions==4.9.0
setuptools>=68.0.0
wheel>=0.40.0
gunicorn==21.2.0 